qa: removed !!; fixed package declaration; Renamed Dkif client and friends;
