+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -30,7 +30,7 @@ import no.nav.syfo.client.wellknown.getWellKnown
 import org.slf4j.LoggerFactory
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
-import no.nav.syfo.client.dkif.DkifClient
+import no.nav.syfo.client.krrproxy.KrrProxyClient
 import no.nav.syfo.oppfolgingsplanmottak.kafka.FollowUpPlanProducer
 import no.nav.syfo.oppfolgingsplanmottak.service.FollowUpPlanSendingService
 
@@ -74,8 +74,8 @@ private fun createApplicationEngineEnvironment(): ApplicationEngineEnvironment {
     val azureAdClient = AzureAdClient(appEnv.auth)
     val isdialogmeldingClient = IsdialogmeldingClient(appEnv.urls, azureAdClient)
     val pdlClient = PdlClient(appEnv.urls, azureAdClient)
-    val dkifClient = DkifClient(appEnv.urls, azureAdClient)
-    val pdfGenClient = OpPdfGenClient(appEnv.urls, appEnv.application, pdlClient, dkifClient)
+    val krrProxyClient = KrrProxyClient(appEnv.urls, azureAdClient)
+    val pdfGenClient = OpPdfGenClient(appEnv.urls, appEnv.application, pdlClient, krrProxyClient)
     val navLpsProducer = AltinnOppfolgingsplanProducer(appEnv.kafka)
     val dokarkivClient = DokarkivClient(appEnv.urls, azureAdClient)
