+++ src/main/kotlin/no/nav/syfo/application/environment/UrlEnv.kt
@@ -10,6 +10,6 @@ data class UrlEnv(
     val dokarkivScope: String,
     val istilgangskontrollUrl: String,
     val istilgangskontrollClientId: String,
-    val dkifUrl: String,
-    val dkifScope: String,
+    val krrProxyUrl: String,
+    val krrProxyScope: String,
 )
