+++ src/main/kotlin/no/nav/syfo/application/environment/Utils.kt
@@ -75,8 +75,8 @@ fun getEnv(): ApplicationEnvironment {
             dokarkivScope = getEnvVar("DOKARKIV_SCOPE"),
             istilgangskontrollUrl = getEnvVar("ISTILGANGSKONTROLL_URL"),
             istilgangskontrollClientId = getEnvVar("ISTILGANGSKONTROLL_CLIENT_ID"),
-            dkifUrl = getEnvVar("DKIF_URL"),
-            dkifScope = getEnvVar("DKIF_SCOPE"),
+            krrProxyUrl = getEnvVar(" KRR_PROXY_URL"),
+            krrProxyScope = getEnvVar(" KRR_PROXY_SCOPE"),
         ),
         altinnLps = AltinnLpsEnv(
             sendToFastlegeRetryThreshold = getEnvVar("SEND_TO_FASTLEGE_RETRY_THRESHOLD").toInt(),
